AzureFunctions/fa_predictive_maintenance_python/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions>=1.13.0b1
pandas
azure.ai.anomalydetector
azure-storage-blob